requirements.txt
scipy
sklearn
numpy
pandas
python-dateutil
importlib
statsmodels
matplotlib
ipywidgets
pymysql 
pynverse
websocket-client
ta
multiprocess